requirements.txt
amqp==5.0.6
asgiref==3.3.4
billiard==3.6.4.0
celery==5.1.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
Django==3.2.3
django-rest-framework==0.1.0
djangorestframework==3.12.4
gcloud==0.17.0
googleapis-common-protos==1.53.0
gunicorn==20.1.0
httplib2==0.19.1
jws==0.1.3
kombu==5.1.0
numpy==1.20.3
oauth2client==3.0.0
opencv-python-headless==4.5.2.52
prompt-toolkit==3.0.18
protobuf==3.17.1
psycopg2-binary==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycryptodome==3.4.3
pyparsing==2.4.7
Pyrebase==3.0.27
python-jwt==2.0.1
pytz==2021.1
requests==2.11.1
requests-toolbelt==0.7.0
rsa==4.7.2
six==1.16.0
sqlparse==0.4.1
vine==5.0.0
wcwidth==0.2.5
